+++ changelog.md
@@ -3,6 +3,7 @@
 * Bugfix: navigation item wasn't using the sortable scope.
 * Bugfix: make sure content widgets can have camelcased model names.
 * Order redirects based on the times they've been used.
+* Add a catch all route with a basic controller that checks for 404's
 
 
 0.0.11 - 2015-12-05